moved components and examples to /components/react
