+++ src/components/react/components/Marquee.tailwind.txt
