+++ src/components/react/components/RangeSlider.tsx
