+++ src/components/react/examples/AvatarExample.tsx
