+++ src/components/react/examples/BadgeExample.tsx
