+++ src/components/react/examples/InputExample.tsx
