+++ src/components/react/examples/RangeSliderExample.tsx
