+++ src/components/react/examples/TabsExample.tsx
